feat:organización de estructura en DashboardLayout
