+++ src/features/dashboard/types/dashboard.types.ts
@@ -1,0 +1,3 @@
+export interface MedicalHistoryProps {
+  children: React.ReactNode;
+}
